// src/app/shared/global-constants.enum.ts
export enum GlobalConstants {
  MY_ALBUMS_STRING = 'myalbums',
  MY_SONGS_STRING = 'mysongs',
  ALBUM_TITLE_STRING = 'albumTitle',
  SONG_TITLE_STRING = 'songTitle',
  FAVORITE_STRING = 'favorite',
  NOT_FOUND_STRING = 'notfound',
  MY_FAVORITES_STRING = 'myfavorites',
  ALBUM_TYPE_STRING = 'album',
  SONG_TYPE_STRING = 'song',
  ADD_ALBUM_STRING = "addalbum"
}